Add automatic PR checks
This adds automatic checks for each PR:

  * `cargo clippy`
  * `cargo fmt` (checks that `cargo fmt` is a no-op)
  * [`cargo msrv`] (verifies that minimum supported Rust version is what
    is set as `rust-version` in Cargo.toml)
  * `cargo test`

This also sets GitHub’s dependabot to check for updates to the GitHub
actions used monthly.

Currently, the crate works on any version of Rust from 1.56 on up. This
sets `rust-version` in Cargo.toml so that any change to the minimum
supported Rust version (MSRV) will be noticed.

Finally, this removes the rust-toolchain file and adds it to .gitignore.
This allows the crate to be developed on any version of Rust ≥1.56.

Fixes: #17 — Set up continious integration

[`cargo msrv`]: https://github.com/foresterre/cargo-msrv